fix(web): serve a minimal social image when the branded render fails
The Open Graph route now renders the branded image to a buffer inside a try/catch and falls back to a minimal unbranded image (default font, shorter cache) if satori or the embedded fonts throw, so the route can no longer 500 after the GitHub fetch succeeded. The scene markup moves to lib/og/social-image.tsx.
